donation_app/serializers.py
from rest_framework import serializers

from donation_app.models import Donation


class DonationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Donation
        fields = "__all__"
